refactor: change packages to match module code style
