+++ trends/src/main/kotlin/net/thechance/trends/api/controller/TrendsGlobalExceptionHandler.kt
@@ -1,9 +1,9 @@
 package net.thechance.trends.api.controller
 
-import net.thechance.trends.api.controller.exception.InvalidTrendInputException
-import net.thechance.trends.api.controller.exception.ReelNotFoundException
-import net.thechance.trends.api.controller.exception.TrendCategoryNotFoundException
-import net.thechance.trends.api.controller.exception.TrendUserNotFoundException
+import net.thechance.trends.exception.InvalidTrendInputException
+import net.thechance.trends.exception.ReelNotFoundException
+import net.thechance.trends.exception.TrendCategoryNotFoundException
+import net.thechance.trends.exception.TrendUserNotFoundException
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.ExceptionHandler
