+++ trends/src/main/kotlin/net/thechance/trends/service/CategoryService.kt
@@ -1,12 +1,9 @@
 package net.thechance.trends.service
 
-import net.thechance.trends.api.controller.exception.InvalidTrendInputException
-import net.thechance.trends.api.controller.exception.TrendCategoryNotFoundException
 import net.thechance.trends.entity.Category
 import net.thechance.trends.repository.CategoryRepository
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
-import java.util.UUID
 
 @Service
 @Transactional(readOnly = true)
