+++ trends/src/main/kotlin/net/thechance/trends/service/TrendUserService.kt
@@ -1,8 +1,7 @@
 package net.thechance.trends.service
 
-import net.thechance.trends.api.controller.exception.InvalidTrendInputException
-import net.thechance.trends.api.controller.exception.TrendCategoryNotFoundException
-import net.thechance.trends.entity.Category
+import net.thechance.trends.exception.InvalidTrendInputException
+import net.thechance.trends.exception.TrendCategoryNotFoundException
 import net.thechance.trends.entity.TrendUser
 import net.thechance.trends.repository.CategoryRepository
 import net.thechance.trends.repository.TrendUserRepository
